# rest-api/src/main/java/at/fhtw/swkom/paperless/controller/ApiApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package at.fhtw.swkom.paperless.controller;

import at.fhtw.swkom.paperless.services.dto.Document;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-10-16T14:45:41.998108992+02:00[Europe/Vienna]", comments = "Generator version: 7.6.0")
@Validated
@Tag(name = "api", description = "the api API")
public interface ApiApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /api/documents/{id} : Delete a document by id
     *
     * @param id The id of the document (required)
     * @return Document deleted successfully (status code 204)
     *         or Document does not exist with this id! (status code 404)
     */
    @Operation(
        operationId = "deleteDocument",
        summary = "Delete a document by id",
        responses = {
            @ApiResponse(responseCode = "204", description = "Document deleted successfully"),
            @ApiResponse(responseCode = "404", description = "Document does not exist with this id!")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/documents/{id}"
    )
    
    default ResponseEntity<Void> deleteDocument(
        @Parameter(name = "id", description = "The id of the document", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/documents/{id} : Retrieve a document by id
     *
     * @param id The id of the document (required)
     * @return Success (status code 200)
     *         or Document does not exist with this id! (status code 404)
     */
    @Operation(
        operationId = "getDocument",
        summary = "Retrieve a document by id",
        responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Document.class))
            }),
            @ApiResponse(responseCode = "404", description = "Document does not exist with this id!")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/documents/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Document> getDocument(
        @Parameter(name = "id", description = "The id of the document", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"path\", \"author\" : \"author\", \"created\" : \"created\", \"id\" : 0, \"title\" : \"title\", \"content\" : \"content\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/documents : Return a list of documents
     *
     * @return Success (status code 200)
     */
    @Operation(
        operationId = "getDocuments",
        summary = "Return a list of documents",
        responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Document.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/documents",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Document>> getDocuments(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"path\" : \"path\", \"author\" : \"author\", \"created\" : \"created\", \"id\" : 0, \"title\" : \"title\", \"content\" : \"content\" }, { \"path\" : \"path\", \"author\" : \"author\", \"created\" : \"created\", \"id\" : 0, \"title\" : \"title\", \"content\" : \"content\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /api/documents/post_document : Upload a document
     *
     * @param author  (required)
     * @param title  (required)
     * @param file  (required)
     * @return Document successfully uploaded. (status code 200)
     *         or Document upload failed due to bad request! (status code 400)
     */
    @Operation(
        operationId = "postDocument",
        summary = "Upload a document",
        responses = {
            @ApiResponse(responseCode = "200", description = "Document successfully uploaded."),
            @ApiResponse(responseCode = "400", description = "Document upload failed due to bad request!")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/documents/post_document",
        consumes = { "multipart/form-data" }
    )
    
    default ResponseEntity<Void> postDocument(
        @Parameter(name = "author", description = "", required = true) @Valid @RequestParam(value = "author", required = true) String author,
        @Parameter(name = "title", description = "", required = true) @Valid @RequestParam(value = "title", required = true) String title,
        @Parameter(name = "file", description = "", required = true) @RequestPart(value = "file", required = true) MultipartFile file
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/documents/search : Retrieve list of documents which match the search input
     *
     * @return Found documents which match the search input (status code 200)
     *         or No content found (status code 204)
     */
    @Operation(
        operationId = "searchDocumentContent",
        summary = "Retrieve list of documents which match the search input",
        responses = {
            @ApiResponse(responseCode = "200", description = "Found documents which match the search input", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Document.class)))
            }),
            @ApiResponse(responseCode = "204", description = "No content found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/documents/search",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Document>> searchDocumentContent(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"path\" : \"path\", \"author\" : \"author\", \"created\" : \"created\", \"id\" : 0, \"title\" : \"title\", \"content\" : \"content\" }, { \"path\" : \"path\", \"author\" : \"author\", \"created\" : \"created\", \"id\" : 0, \"title\" : \"title\", \"content\" : \"content\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /api/documents/{id} : Updates the Metadata of a document
     *
     * @param id The id of the document (required)
     * @param title  (required)
     * @param author  (optional)
     * @param file  (optional)
     * @return Document does not exist with this id! (status code 404)
     *         or Metadata updated successfully (status code 200)
     */
    @Operation(
        operationId = "updateMetaData",
        summary = "Updates the Metadata of a document",
        responses = {
            @ApiResponse(responseCode = "404", description = "Document does not exist with this id!"),
            @ApiResponse(responseCode = "200", description = "Metadata updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Document.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/documents/{id}",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    default ResponseEntity<Document> updateMetaData(
        @Parameter(name = "id", description = "The id of the document", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "title", description = "", required = true) @Valid @RequestParam(value = "title", required = true) String title,
        @Parameter(name = "author", description = "") @Valid @RequestParam(value = "author", required = false) String author,
        @Parameter(name = "file", description = "") @RequestPart(value = "file", required = false) MultipartFile file
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"path\", \"author\" : \"author\", \"created\" : \"created\", \"id\" : 0, \"title\" : \"title\", \"content\" : \"content\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}